working comments

+++ server/lib/final_project_web/controllers/comment_controller.ex
@@ -3,6 +3,8 @@ defmodule FinalProjectWeb.CommentController do
 
   alias FinalProject.Comments
   alias FinalProject.Comments.Comment
+  alias FinalProjectWeb.SessionController
+
 
   action_fallback FinalProjectWeb.FallbackController
 
@@ -11,12 +13,21 @@ defmodule FinalProjectWeb.CommentController do
     render(conn, "index.json", comments: comments)
   end
 
-  def create(conn, %{"comment" => comment_params}) do
-    with {:ok, %Comment{} = comment} <- Comments.create_comment(comment_params) do
+  def create(conn, %{"comment" => comment_params, "session" => session}) do
+    if SessionController.authorized?(conn, session["user_id"], session["token"]) do
+      with {:ok, %Comment{} = _comment} <- Comments.create_comment(comment_params) do
+        posts = FinalProject.Posts.get_recent_posts()
+        FinalProjectWeb.Endpoint.broadcast("feed:", "feed/update", FinalProjectWeb.PostView.render("index.json", posts: posts))
+
+        conn
+        |> send_resp(:created, "Success")
+      end
+    else
       conn
-      |> put_status(:created)
-      |> put_resp_header("location", Routes.comment_path(conn, :show, comment))
-      |> render("show.json", comment: comment)
+      |> put_resp_header(
+        "content-type",
+        "application/json; charset=UTF-8")
+      |> send_resp(:unauthorized, Jason.encode!(%{error: "Unauthorized"}))
     end
   end
 
@@ -33,11 +44,29 @@ defmodule FinalProjectWeb.CommentController do
     end
   end
 
-  def delete(conn, %{"id" => id}) do
+  def delete(conn, %{"id" => id, "session" => session}) do
     comment = Comments.get_comment!(id)
-
-    with {:ok, %Comment{}} <- Comments.delete_comment(comment) do
-      send_resp(conn, :no_content, "")
+    user_id = session["user_id"]
+    if SessionController.authorized?(conn, user_id, session["token"]) do
+      if (user_id == comment.user.id) || (user_id == comment.post.user.id) do
+        with {:ok, %Comment{}} <- Comments.delete_comment(comment) do
+          posts = FinalProject.Posts.get_recent_posts()
+          FinalProjectWeb.Endpoint.broadcast("feed:", "feed/update", FinalProjectWeb.PostView.render("index.json", posts: posts))
+          send_resp(conn, :no_content, "")
+        end
+      else
+        conn
+        |> put_resp_header(
+          "content-type",
+          "application/json; charset=UTF-8")
+        |> send_resp(:unauthorized, Jason.encode!(%{error: "Unauthorized"}))
+      end
+    else
+      conn
+      |> put_resp_header(
+        "content-type",
+        "application/json; charset=UTF-8")
+      |> send_resp(:unauthorized, Jason.encode!(%{error: "Unauthorized"}))
     end
   end
 end
